model_training/2d_landmarks_model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnxruntime-gpu>=1.15.0
onnx>=1.14.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
mxnet==1.9.1
insightface>=0.7.3
scikit-learn>=1.3.0
psutil
onnx-simplifier
onnxscript
gdown
